requirements.txt
beautifulsoup4==4.6.0
celery==4.1.1
cryptography==2.4.2
Flask==1.0.2
Flask-Migrate==1.8.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1 
Flask-JWT-Extended==3.12.0
google-cloud-storage==1.16.1
gevent==1.3.7
gunicorn==19.7.1
idna==2.7
imagehash==4.0
iso8601==0.1.12
jsonschema==2.6.0
kombu==4.5.0
pillow==6.2.1
psycopg2-binary==2.7.5
pyOpenSSL==18.0.0
pypng==0.0.20
pyqrcode==1.2.1
redis==3.3.5
requests>=2.0.0
six==1.11.0
Wand==0.5.8

-e git+https://github.com/NewAcropolis/na-common.git@afd9e32dfe2d6e7c6c72f61dee417cf85a1b625e#egg=na-common
